code changes and test cases

+++ python/cudf/cudf/core/series.py
@@ -3778,7 +3778,11 @@ class Series(Frame, Serializable):
             data = _format_stats_values(data)
 
             return Series(
-                data=data, index=names, nan_as_null=False, name=self.name
+                data=data,
+                index=names,
+                nan_as_null=False,
+                name=self.name,
+                dtype="float",
             )
 
         def describe_categorical(self):
